[add] IdentifierKeyを追加
変数許容文字列
